refactor: refatora controller e rotas
