+++ src/controller/user.js
@@ -1,50 +1,41 @@
-const db = require('../database/dao');
-const bcrypt = require('bcryptjs');
+const user = require('../models/userModel')
 
-
-const insertUser = (user) => {
-    user.password = bcrypt.hashSync(user.password, 10);
-    db.insert('users', user, (result) =>{
-        if(result.error){
-            throw result.error;
+module.exports = {
+    login(req, res) {
+        let usr = req.body;
+        if(req.session.user){
+            res.redirect('/');
+        } else if(!user.findOne(usr.username)){
+            res.status(401).send({error: 'User not found'});
+        } else if(user.find(usr)){
+            req.session.user = usr.username
+            res.status(200).send("OK");
+        } else {
+            res.status(403).send({error: 'Not authorized'});
         }
-    });
-}
-
-const deleteUser = (user) => {
-    const SQL_DELETE = `DELETE FROM users WHERE username = "${user.username}";`
-    db.run(SQL_DELETE,(result) => {
-        if(result.error){
-            throw result.error;
+    },
+    create (req, res){
+        let usr = {
+            username: req.body.username,
+            password: req.body.password
         }
-    });
+        if(usr.username.length < 4){
+            res.status(400).send({error: "Username is too small"});
+        } else if(usr.password.length < 4) {
+            res.status(400).send({error: "Password is too small"})
+        } else if(user.findOne(usr.username)){
+            res.status(400).send({error: "Username already exists"});
+        } else {
+            user.insert(usr);
+            req.session.user = usr.username
+            console.log(usr)
+            res.status(200).send("OK");
+        }
+    },
+    verifyUsername(req, res){
+        usr = req.body;
+        if(user.findOne(usr.username)){
+            res.status(400).send({error: "Username already exists"});
+        }
+    }
 }
-
-const updateUser = (user) => {
-    db.update('users', user, `username = "${user.username}"`);
-}
-
-const findOne = (username) => {
-    let SQL_SELECT_ONE = `SELECT username FROM users
-                                WHERE username = "${username}";`;
-    return result = db.run(SQL_SELECT_ONE)[0];
-}
-
-const findUser = (user) => {
-    let SQL_SELECT_USER = `SELECT * FROM users
-                                WHERE username = "${user.username}";`;
-    let result = db.run(SQL_SELECT_USER)[0];
-    return result && bcrypt.compareSync(user.password, result.password);
-}
-
-const listUsers = () => {
-    let SQL_QUERY = `SELECT username FROM users;`;
-    return db.run(SQL_QUERY);
-}
-
-exports.insert = insertUser;
-exports.delete = deleteUser;
-exports.update = updateUser;
-exports.find = findUser;
-exports.findOne = findOne;
-exports.list = listUsers;
