+++ src/routes/login.js
@@ -6,21 +6,9 @@ router.get('/', (req, res) => {
     if(req.session.user){
         return res.redirect('/');
     }
-    res.render('login.ejs');
+    return res.render('login.ejs');
 });
 
-router.post('/', (req, res) => {
-    let usr = req.body;
-    if(req.session.user){
-        res.redirect('/');
-    } else if(!user.findOne(usr.username)){
-        res.status(401).send({error: 'User not found'});
-    } else if(user.find(usr)){
-        req.session.user = usr.username
-        res.status(200).send("OK");
-    } else {
-        res.status(403).send({error: 'Not authorized'});
-    }
-});
+router.post('/', user.login);
 
 module.exports = router;
