+++ src/routes/register.js
@@ -6,30 +6,8 @@ router.get('/', (req, res) => {
     res.render('register.ejs');
 });
 
-router.post('/', (req, res) => {
-    let usr = {
-        username: req.body.username,
-        password: req.body.password
-    }
-    if(usr.username.length < 4){
-        res.status(400).send({error: "Username is very small"});
-    } else if(usr.password.length < 4) {
-        res.status(400).send({error: "Password is very small"}) 
-    } else if(user.findOne(usr.username)){
-        res.status(400).send({error: "Username already exists"});
-    } else {
-        user.insert(usr);
-        req.session.user = usr.username
-        console.log(usr)
-        res.status(200).send("OK");
-    }
-});
+router.post('/', user.create);
 
-router.post('/username', (req, res) => {
-    usr = req.body;
-    if(user.findOne(usr.username)){
-        res.status(400).send({error: "Username already exists"});
-    }
-})
+router.post('/username', user.verifyUsername)
 
 module.exports = router;
